# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Medioa REST API (with gin-gonic).",
        "title": "Medioa API",
        "contact": {
            "name": "Vũ Kỳ",
            "url": "github.com/vukyn",
            "email": "vukynpro@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/storage/download/request/{file_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get download url for private media with download password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Request download private media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secret",
                        "name": "secret",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.RequestDownloadResponse"
                        }
                    }
                }
            }
        },
        "/storage/download/{file_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Download media file (images, videos, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Download media (public/private)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secret",
                        "name": "secret",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "silent response",
                        "name": "silent",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.DownloadResponse"
                        }
                    }
                }
            }
        },
        "/storage/secret": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new secret for upload media",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Create new secret",
                "parameters": [
                    {
                        "description": "create secret request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.CreateSecretRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.CreateSecretResponse"
                        }
                    }
                }
            }
        },
        "/storage/secret/pin": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Reset pin code for secret",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Reset pin code",
                "parameters": [
                    {
                        "description": "reset pin request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.ResetPinCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/storage/secret/retrieve": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve secret with new access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Retrieve secret",
                "parameters": [
                    {
                        "description": "retrieve secrect request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.RetrieveSecretRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.RetrieveSecretResponse"
                        }
                    }
                }
            }
        },
        "/storage/secret/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload media file (images, videos, etc.)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Upload media with secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "secret",
                        "name": "secret",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "binary file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file name",
                        "name": "file_name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.UploadResponse"
                        }
                    }
                }
            }
        },
        "/storage/secret/upload/commit": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Commit all chunks to complete upload media file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Commit upload media chunk with secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "secret",
                        "name": "secret",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "commit chunk request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.CommitChunkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.CommitChunkResponse"
                        }
                    }
                }
            }
        },
        "/storage/secret/upload/stage": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload media file (images, videos, etc.)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Upload media by chunk with secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "secret",
                        "name": "secret",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "binary chunk",
                        "name": "chunk",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "chunk index",
                        "name": "chunk_index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "total chunk",
                        "name": "total_chunks",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "file_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "file name",
                        "name": "file_name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.UploadChunkResponse"
                        }
                    }
                }
            }
        },
        "/storage/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload media file (images, videos, etc.)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Upload media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "binary file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file name",
                        "name": "file_name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.UploadResponse"
                        }
                    }
                }
            }
        },
        "/storage/upload/commit": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Commit all chunks to complete upload media file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Commit upload media chunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "description": "commit chunk request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.CommitChunkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.CommitChunkResponse"
                        }
                    }
                }
            }
        },
        "/storage/upload/stage": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload media file (images, videos, etc.)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Upload media by chunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "binary chunk",
                        "name": "chunk",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "chunk index",
                        "name": "chunk_index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "total chunk",
                        "name": "total_chunks",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "file id",
                        "name": "file_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "file name",
                        "name": "file_name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/medioa_internal_storage_models.UploadChunkResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "medioa_internal_storage_models.CreateSecretRequest": {
            "type": "object",
            "properties": {
                "master_key": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "pin_code": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "medioa_internal_storage_models.CreateSecretResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "medioa_internal_storage_models.DownloadResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "medioa_internal_storage_models.RequestDownloadResponse": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "medioa_internal_storage_models.ResetPinCodeRequest": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "new_pin_code": {
                    "type": "string"
                }
            }
        },
        "medioa_internal_storage_models.RetrieveSecretRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}